+++ src/main/kotlin/org/rescado/server/controller/dto/req/AuthWithTokenDTO.kt
@@ -16,10 +16,10 @@ data class AuthWithTokenDTO(
     @get:Pattern(message = "{Pattern.AuthWithTokenDTO.refreshToken}", regexp = "^[\\w]{8}-[\\w]{4}-[\\w]{4}-[\\w]{4}-[\\w]{12}$")
     val token: String,
 
-    @get:InRange(min = -90f, max = 90f, message = "{InRange.AuthWithTokenDTO.latitude}")
-    override val latitude: Float?,
+    @get:InRange(min = -90.0, max = 90.0, message = "{InRange.AuthWithTokenDTO.latitude}")
+    override val latitude: Double?,
 
-    @get:InRange(min = -180f, max = 180f, message = "{InRange.AuthWithTokenDTO.longitude}")
-    override val longitude: Float?,
+    @get:InRange(min = -180.0, max = 180.0, message = "{InRange.AuthWithTokenDTO.longitude}")
+    override val longitude: Double?,
 
 ) : AuthDTO(latitude, longitude)
