+++ src/main/kotlin/org/rescado/server/controller/dto/validation/InRange.kt
@@ -13,21 +13,21 @@ annotation class InRange(
     val groups: Array<KClass<out Any>> = [],
     val payload: Array<KClass<out Any>> = [],
 
-    val min: Float,
-    val max: Float,
+    val min: Double,
+    val max: Double,
 )
 
-class InRangeValidator : ConstraintValidator<InRange, Float> {
+class InRangeValidator : ConstraintValidator<InRange, Double> {
 
-    private var min = 0.0f
-    private var max = 0.0f
+    private var min = 0.0
+    private var max = 0.0
 
     override fun initialize(constraintAnnotation: InRange) {
         this.min = constraintAnnotation.min
         this.max = constraintAnnotation.max
     }
 
-    override fun isValid(value: Float?, context: ConstraintValidatorContext?): Boolean {
+    override fun isValid(value: Double?, context: ConstraintValidatorContext?): Boolean {
         return if (value == null) true else value in this.min..this.max
     }
 }
